[Silver I] Title: 오르막 수, Time: 124 ms, Memory: 14244 KB -BaekjoonHub

+++ 백준/Silver/11057. 오르막 수/오르막 수.java
@@ -9,25 +9,25 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
 
-        int[][] dp = new int[N + 1][10];
+        long[][] dp = new long[N + 1][10];
         for (int j = 0; j < 10; j++) {
             dp[1][j] = 1;
         }
 
         for (int i = 2; i <= N; i++) {
-            long sum = 0; // Change data type to long
+            long sum = 0;
             for (int j = 0; j < 10; j++) {
                 if (j == 0) {
                     dp[i][0] = dp[i - 1][0];
                 }
                 sum = (sum + dp[i - 1][j]) % mod;
-                dp[i][j] = (int) sum; // Convert back to int for dp array
+                dp[i][j] = sum; // Convert back to int for dp array
             }
         }
 
-        int result = 0;
+        long result = 0;
         for (int j = 0; j < 10; j++) {
-            result = (result + dp[N][j]) % (int) mod; // Convert back to int for result
+            result =  (result + dp[N][j]) %  mod; // Convert back to int for result
         }
 
         System.out.println(result);
